requirements.txt
spotipy==2.4.4
requests==2.18.1
Pillow==4.2.0
